refactor(sidebar): split the scope into six stores and unify the drag CSS variables
Completes the two steps the previous PR left out.

`sidebar-scope.ts` becomes composition: six factories under `sidebar/stores/`,
each built in dependency order because a store can only be handed what already
exists. `createSidebarDragEngine` drops from 776 to 455 lines - it keeps the
imperative pipeline and is handed only the drag store's writers, so reading a
one-flush-stale `drag.source` is now a type error rather than a code review
note. `commitSidebarDrop` takes its source as a parameter for the same reason.

`announceError` folds seven copies of the same `instanceof Error` unwrapping.

The four drag offset transports become one pair of custom properties,
`--sidebar-drag-x/y`, plus one `.sidebar-drag-shifting` class. Custom properties
inherit, so every shiftable row declares the pair on itself: an agent row inside
a shifting section would otherwise inherit its offset and move twice.

Verified in Storybook against a live drag of all four regions - the section
shifts -266px while its three agent rows stay at 0px.

Surfaces: desktop renderer only.
